# perf: Take advantage of fast networks for file transfers