@JFreegman 2020-12-24
Tests need to be conducted in real world conditions with clients that have high latency/low bandwidth connections, and on clients other than Toxic, as most clients don't have file resuming which means file transfers get reset on connection breakage. This used to be a major issue that rendered file transfers useless for many people, so we need to make sure that we don't re-introduce that problem.

This PR needs to be run through astyle as well.